changed icons style moved copyFolder function to helper.js
