+++ TrafficPrediction/my_modules/utils/helper.js
@@ -1,3 +1,7 @@
+var logger = require("./logger/logger.js");
+var qm = require("qminer");
+var path = require('path');
+
 // Dummy model, used for feature extractor
 // .setVal() -- sets internal value
 // .getVal() -- returns internal value
@@ -23,4 +27,27 @@ exports.newDummyModel = function () {
         }
     }
     return new dummyModel();
+}
+
+exports.copyFolder = function (inFolder, outFolder) {
+    logger.debug("Copying ." + path.basename(inFolder) + " to ." + path.basename(outFolder) + " folder...");
+    
+    // read all files in inFolder
+    var files = qm.fs.listFile(inFolder, null, true);
+    
+    // create outFolder if it doesnet exist
+    if (!qm.fs.exists(outFolder)) qm.fs.mkdir(outFolder);
+    
+    // copy files from inFloder to outFolder
+    files.forEach(function (file) {
+        var source = path.normalize(file);
+        var dest = source.replace(inFolder, outFolder);
+        
+        // copy file one by one
+        if (qm.fs.exists(file)) {
+            if (!qm.fs.exists(path.dirname(dest))) qm.fs.mkdir(path.dirname(dest));
+            qm.fs.copy(source, dest);
+        }
+    });
+    logger.debug(files.length + " files copied.");
 }
