--- TrafficPrediction/public/css/css.css
@@ -1,151 +1,0 @@
-﻿html, body {
-  height: 100%;
-  width: 100%;
-  padding: 0;
-  margin: 0;
-}
-
-.page-wrapper {
-  height: 100%;
-  position: relative;
-}
-
-.google-maps {
-    position: absolute;
-    top: 40px;
-    left: 0;
-    width: 100% !important;
-    height: 100% !important;
-    margin-bottom:-43px;
-    z-index: 0;
-}
-
-.topBar{
-   position:fixed;
-   top:0;
-   left:0;
-   width:100%;
-   height:40px;
-   background-color:white;
-   border-bottom: solid #D11042;
-   display:inline-block;
-}
-
-.topElement{
-   display:inline-block;
-   vertical-align: top;
-   height: 100%;
-}
-
-.prediction-picker{
-   display:inline-block;
-   margin-right:5px;
-   margin-top:6px;
-   padding-left: 10px;
-   width:220px;
-}
-
-#amount{
-   display:inline-block;
-   margin-right:5px;
-   margin-top:6px;
-   width:30px;
-   font-size: large;  
-}
-
-.right-content{
-  position: fixed;
-  top: 40px;
-  right: 0px;
-  bottom: 0px;
-  width:500px;
-  padding: 10px;
-  border-left: solid #D11042;
-  background-color:white;
-}
- 
- #green{
-   height: 20px !important;
- }
- 
-.right-slider{
-   margin-top:13px;
-   margin-left:5px;
-   margin-right:5px;
-   width:450px;
-}
-
-#slider-range-green.ui-slider .ui-slider-range {
-    background: #12BA1F;
-}
-  
-#slider-range-yellow.ui-slider .ui-slider-range {
-   background: #EEDD00;
-}
-
-#slider-range-orange.ui-slider .ui-slider-range {
-   background: #FFBD00;
-}
-
-.tresholds {
-  margin-top: 10px !important;
-}
-
-.topSlider{
-   margin-top:13px;
-   margin-left:5px;
-   margin-right:5px;
-   width:500px;
-}
-
-  
-#alternating-slider {
-  height: 500px;
-  margin-left: 30px;
-}
-  
-#treshold-options-button {
-  position: fixed;
-  right: 5px;  
-  top: 5px;
-	background:-webkit-gradient(linear, left top, left bottom, color-stop(0.05, #f2fefd), color-stop(1, #f0f0f0));
-	background:-moz-linear-gradient(top, #f2fefd 5%, #f0f0f0 100%);
-	background:-webkit-linear-gradient(top, #f2fefd 5%, #f0f0f0 100%);
-	background:-o-linear-gradient(top, #f2fefd 5%, #f0f0f0 100%);
-	background:-ms-linear-gradient(top, #f2fefd 5%, #f0f0f0 100%);
-	background:linear-gradient(to bottom, #f2fefd 5%, #f0f0f0 100%);
-	filter:progid:DXImageTransform.Microsoft.gradient(startColorstr='#f2fefd', endColorstr='#f0f0f0',GradientType=0);
-	background-color:#f2fefd;
-	-moz-border-radius:18px;
-	-webkit-border-radius:18px;
-	border-radius:18px;
-	border:1px solid #f5e8e7;
-	display:inline-block;
-	cursor:pointer;
-	color:#D11042;
-	padding:3px 25px;
-	text-decoration:none;
-	text-shadow:0px 1px 0px #D8ACCF;
-  /*font-family: "Helvetica Neue",Helvetica,Arial,sans-serif;*/
-}
-
-#treshold-options-button:hover {
-	background:-webkit-gradient(linear, left top, left bottom, color-stop(0.05, #f0f0f0), color-stop(1, #f2fefd));
-	background:-moz-linear-gradient(top, #f0f0f0 5%, #f2fefd 100%);
-	background:-webkit-linear-gradient(top, #f0f0f0 5%, #f2fefd 100%);
-	background:-o-linear-gradient(top, #f0f0f0 5%, #f2fefd 100%);
-	background:-ms-linear-gradient(top, #f0f0f0 5%, #f2fefd 100%);
-	background:linear-gradient(to bottom, #f0f0f0 5%, #f2fefd 100%);
-	filter:progid:DXImageTransform.Microsoft.gradient(startColorstr='#f0f0f0', endColorstr='#f2fefd',GradientType=0);
-	background-color:#f0f0f0;
-}
-
-#treshold-options-button:active {
-  position: fixed;
-  right: 7px;  
-  top: 3px;
-}
-
-#treshold-options-button:focus {
-  outline:0;
-}
